Add sticky side headings
Add h1 level headings to label professional and personal projects. Make
it such that the headings are sideways reading botton to top and that
the are sticky so that while one is scrolling the professional section
the professional label is there which then gets overtaken by the
personal heading when the reader reaches the personal section.