Fix sliding behaviour of sideways headings
Seperate out the Professional and Personal tables into seperate elements
so that the Professional heading slides up once the end of the
professional work is reached and the Personal heading slides in its
place.

Also add a high z-index to the nav bar so that the haeadings slide
behind it not in front of it.